Settings now saves trips to state as well as Database
