+++ server/controllers/roadtripController.js
@@ -4,7 +4,7 @@ const roadtripController = {};
 
 roadtripController.createRoadtrip = async (req, res, next) => {
   console.log('Made it to RoadtripController!');
-  console.log(req);
+  // console.log(req);
   const roadtripName = req.body.roadtripName;
   const creator = req.cookies.sessionId;
   const tripData = req.body.tripData;
